+++ src/interfaces/graphQL/singleDataSource.ts
@@ -3,7 +3,7 @@
  */
 import {DimensionFilter} from "./dimensionFilter";
 import {Metric} from "./metric";
-import {DataSourceType} from "../../models/types";
+import {DataSourceType} from "../../types/types";
 import {MetricFilter} from "./metricFilter";
 import {VersionFilter} from "./versionFilter";
 
@@ -14,5 +14,5 @@ export interface SingleDataSource {
     dimensions: DimensionFilter[];      // набор фильтров по полям (применим только для SINGLE)
     metric: Metric;                     // метрика которую надо расчитать
     metricFilters: MetricFilter[];      // Дополнительные условия для фильтрации по метрикам
-    versionFilter: VersionFilter;       // Фильтр версий
+    versionFilter?: VersionFilter;      // Фильтр версий
 }